Refactored components and improved visual aspects
A lot of similar components, like variations of buttons, were removed.
Instead, a simple generic component was added on it's place with a
"type" mandatory props, required by TypeScript. Based in this "type"
prop, which is a union of strings, React will render dynamically a lot
of types of buttons and require a custom set of another types as well.